You will need to fix all rust linting errors.
please make it so that when I am hovering over an attribute of a map type a-la slime.name I get the type of that attribute

Also map's should have to be initialized with all the keys' types we expect, OR a single specific type

if possible can we also make the typing intellisense for things like bag show the full typing from the annotation e.g. not just bag is a list but that it's a list of strings?